cambio validacion

+++ app/Http/Controllers/Auth/AuthenticatedSessionController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\Auth\LoginRequest;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Validator;
 
 
 class AuthenticatedSessionController extends Controller
@@ -17,9 +18,16 @@ class AuthenticatedSessionController extends Controller
     public function store(LoginRequest $request)
     {
         $request->session()->regenerate();
+        $validator = Validator::make($request->all(), [
+            'email' => 'exists:usuarios,email',
+        ], ['email.exists'=> 'El usuario no existe']);
+
+        if($validator->fails()){
+            return response()->json(['message'=> $validator->errors()->first()], 422);
+        }
 
         if (!Auth::attempt($request->only('email', 'password'))) {
-            return response()->json(['message' => 'usuario o contraseña inválidos'], 422);
+            return response()->json(['message' => 'contraseña incorrecta'], 422);
         }
         Auth::user()->tokens()->delete();
         $token = Auth::user()->createToken('token', ['usuario-registrado']);
